+++ src/System.Net.NameResolution/src/System/Net/NameResolutionPal.Windows.Uap.cs
@@ -1,0 +1,15 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Net
+{
+    internal static partial class NameResolutionPal
+    {
+
+        private static bool GetAddrInfoExSupportsOverlapped()
+        {
+            return false;
+        }
+    }
+}
